Introduce the 4th version, the unlimited number of parameters.
The intention of this version is to improve the flexibility of the
custom function.
Since the old version we allow only up to 2 parameters to both function
and condition
that might not enough for custom ones.

Since the version 4, any user can define the unlimited number of
parameters in the metadata file,
then refer them using the parameter name in the configuration. This
could provide the more
flexible configuration since users don't need to config the pre-function
in order.

The readme.md was rewriten and separated the old one to readme_v3.md so
that user could catch up if they are using the old version and tring to
upgrade to the newer version.

Introduce new meaningful exceptions

In order to separate the error from user configuration, or the
framework implementation, there are some new execeptions have been
introduced including
1. InvalidConfigurationException
2. NoSuchExecutableMethodException
3. NoSuchDefinitionException